+++ src/xnatuploader/xnatuploader.py
@@ -16,6 +16,7 @@ __version__ = version("xnatuploader")
 from openpyxl import load_workbook
 
 from xnatuploader.matcher import Matcher
+from xnatuploader.dicoms import dicom_extractor, XNATFileMatch, SPREADSHEET_FIELDS
 from xnatuploader.workbook import new_workbook, add_filesheet, load_config
 from xnatuploader.upload import Upload, trigger_pipelines
 
@@ -46,7 +47,7 @@ def scan(matcher, root, spreadsheet, include_unmatched=True, debug=False):
 
     If the debug flag is true, only try to match DEBUG_MAX files
     ---
-    matcher: a PathMatcher
+    matcher: a Matcher
     root: pathlib.Path
     spreadsheet: pathlib.Path
     include_unmatched: boolean
@@ -55,34 +56,48 @@ def scan(matcher, root, spreadsheet, include_unmatched=True, debug=False):
     logger.info(f"Loading {spreadsheet}")
     wb = load_workbook(spreadsheet)
     ws = add_filesheet(wb, matcher, debug)  # keeps old sheets if debug=True
-    matched = 0
-    unmatched = 0
     logger.info("Preparing file list")
-    files = sorted(
+    filepaths = sorted(
         [f for f in root.glob("**/*") if f.is_file() and f.name not in IGNORE_FILES]
     )
+    files = []
+    unmatched = []
     if debug:
-        files = files[:DEBUG_MAX]
+        filepaths = filepaths[:DEBUG_MAX]
     logger.info(f"Scanning directory {root}")
-    for file in tqdm(files):
-        if file.is_file():
-            logger.debug(f"Scanning {file}")
-            filematch = matcher.match(root, file)
-            if filematch.success:
-                matched += 1
-                logger.debug(f"Matched {filematch.file}")
-                ws.append(filematch.columns)
+    for filepath in tqdm(filepaths):
+        if filepath.is_file():
+            logger.debug(f"Scanning {filepath}")
+            file = matcher.match(root, filepath)
+            if file.success:
+                logger.debug(f"Matched {file.file}")
+                files.append(file)
             else:
                 if include_unmatched:
-                    unmatched += 1
-                    filematch.load_dicom()
-                    ws.append(filematch.columns)
+                    file.load_dicom()
+                    unmatched.append(file)
+
+    skips, uploads = collate_uploads(files)
+
+    ns = len(uploads)
+    nm = len(files)
+    num = len(unmatched)
+
     if include_unmatched:
         logger.info(
-            f"Saved {matched} matching files and {unmatched} non-matching files to {spreadsheet}"
+            f"Saving {ns} scans with {nm} matching files and {num} non-matching files to {spreadsheet}"
         )
     else:
-        logger.info(f"Saved {matched} matching files to {spreadsheet}")
+        logger.info(f"Saving {ns} scans with {nm} matching files to {spreadsheet}")
+
+    for session_scan, upload in tqdm(uploads.items(), desc="Scans"):
+        for file in upload.files:
+            ws.append(file.columns)
+
+    if include_unmatched:
+        for file in unmatched:
+            ws.append(file.columns)
+
     wb.save(spreadsheet)
 
 
@@ -126,7 +141,7 @@ def upload(
         else:
             matchfile = matcher.from_spreadsheet(row)
             files.append(matchfile)
-    skip, uploads = collate_uploads(project, files)
+    skip, uploads = collate_uploads(files)
     csvout = get_csv_filename(spreadsheet)
     if test:
         dry_run(uploads)
@@ -241,14 +256,14 @@ The results are available as a CSV file: {csvout}
         )
 
 
-def collate_uploads(project_id, files):
+def collate_uploads(files):
     """
     Takes a list of files and collates them by subject (patient), visit
-    index (starting from the earliest) and scan type, returning a list of
-    files which have skipped or already uploaded and a dictionary
-    of Uploads keyed by {session_label}_{scan}
+    index (starting from the earliest), scan type, and (optionally) scan_id,
+    returning a list of files which have skipped or already uploaded and a dictionary
+    of Uploads keyed by {session_label}_{scan}_{scan_id}
+
     ---
-    project_id: str
     files: list of FileMatch
 
     returns: tuple of ( list of FileMatch, dict of str: Upload )
@@ -264,9 +279,9 @@ def collate_uploads(project_id, files):
                 logger.debug(f"skipping file already uploaded {file.file}")
                 skip.append(file)
             else:
-                if file.subject not in subjects:
-                    subjects[file.subject] = []
-                subjects[file.subject].append(file)
+                if file["Subject"] not in subjects:
+                    subjects[file["Subject"]] = []
+                subjects[file["Subject"]].append(file)
     uploads = {}
     for subject_id, files in subjects.items():
         dates = sorted(set([file.study_date for file in files]))
@@ -275,19 +290,21 @@ def collate_uploads(project_id, files):
         for file in files:
             visit = visits[file.study_date]
             modality = file.modality
-            session_label = f"{subject_id}_{modality}{visit}"
+            scan_id = file.series_number
+            session_label = f"{subject_id}_{modality}{visit}_{scan_id}"
             file.session_label = session_label
             scan_type = clean_datasets[file.dataset]
             session_scan = f"{session_label}:{scan_type}"
             if session_scan not in uploads:
                 uploads[session_scan] = Upload(
-                    session_label,
-                    subject_id,
-                    file.study_date,
-                    modality,
-                    scan_type,
-                    file.manufacturer,
-                    file.model,
+                    session_label=session_label,
+                    subject=subject_id,
+                    date=file.study_date,
+                    modality=modality,
+                    series_number=scan_id,
+                    scan_type=scan_type,
+                    manufacturer=file.manufacturer,
+                    model=file.model,
                 )
             uploads[session_scan].add_file(file)
     return skip, uploads
@@ -466,7 +483,14 @@ debug messages
 
     config = load_config(args.spreadsheet)
 
-    matcher = Matcher(config, loglevel=loglevel)
+    matcher = Matcher(
+        patterns=config["paths"],
+        mappings=config["mappings"],
+        fields=SPREADSHEET_FIELDS,
+        file_extractor=dicom_extractor,
+        match_class=XNATFileMatch,
+        loglevel=loglevel,
+    )
 
     if args.operation == "scan":
         scan(