+++ tests/test_scan.py
@@ -4,6 +4,7 @@ from openpyxl import load_workbook
 from pathlib import Path
 
 from xnatuploader.matcher import Matcher
+from xnatuploader.dicoms import dicom_extractor, XNATFileMatch, SPREADSHEET_FIELDS
 from xnatuploader.xnatuploader import scan, collate_uploads
 from xnatuploader.workbook import load_config, new_workbook
 
@@ -21,7 +22,13 @@ def assert_worksheets_equal(expect, got):
 def test_scan(tmp_path, test_files):
     fileset = test_files["basic"]
     config = load_config(fileset["config_excel"])
-    matcher = Matcher(config)
+    matcher = Matcher(
+        config["paths"],
+        config["mappings"],
+        SPREADSHEET_FIELDS,
+        dicom_extractor,
+        XNATFileMatch,
+    )
     scanned = tmp_path / "scanned.xlsx"
     new_workbook(scanned)
     scan(matcher, Path(fileset["dir"]), scanned, include_unmatched=True)
@@ -34,11 +41,16 @@ def test_scan(tmp_path, test_files):
 def test_collation(tmp_path, test_files, uploads_dict):
     fileset = test_files["basic"]
     config = load_config(fileset["config_excel"])
-    matcher = Matcher(config)
+    matcher = Matcher(
+        config["paths"],
+        config["mappings"],
+        SPREADSHEET_FIELDS,
+        dicom_extractor,
+        XNATFileMatch,
+    )
     log = tmp_path / "log.xlsx"
     new_workbook(log)
     scan(matcher, Path(fileset["dir"]), log)
-    project_id = uploads_dict["project"]
     wb = load_workbook(log)
     ws = wb["Files"]
     header = True
@@ -49,7 +61,7 @@ def test_collation(tmp_path, test_files, uploads_dict):
         else:
             matchfile = matcher.from_spreadsheet(row)
             files.append(matchfile)
-    skipped, uploads = collate_uploads(project_id, files)
+    skipped, uploads = collate_uploads(files)
     for session_scan, upload in uploads.items():
         uploads[session_scan] = [f.file for f in uploads[session_scan].files]
     assert uploads == uploads_dict["uploads"]
@@ -61,12 +73,17 @@ def test_sanitisation_collisions(tmp_path, test_files, sanitised_dict):
     config_file = fileset["config"]
     with open(config_file, "r") as fh:
         config = json.load(fh)
-    matcher = Matcher(config)
+    matcher = Matcher(
+        config["paths"],
+        config["mappings"],
+        SPREADSHEET_FIELDS,
+        dicom_extractor,
+        XNATFileMatch,
+    )
     logger.warning(f"Testing santisation in {fileset}")
     log = tmp_path / "log.xlsx"
     new_workbook(log)
     scan(matcher, Path(fileset["dir"]), log)
-    project_id = sanitised_dict["project"]
     wb = load_workbook(log)
     ws = wb["Files"]
     header = True
@@ -78,7 +95,7 @@ def test_sanitisation_collisions(tmp_path, test_files, sanitised_dict):
             matchfile = matcher.from_spreadsheet(row)
             files.append(matchfile)
     logger.warning(f"File list = {files}")
-    skipped, uploads = collate_uploads(project_id, files)
+    skipped, uploads = collate_uploads(files)
     for session_scan, upload in uploads.items():
         uploads[session_scan] = [f.file for f in uploads[session_scan].files]
     assert uploads == sanitised_dict["uploads"]
@@ -88,11 +105,16 @@ def test_sanitisation_collisions(tmp_path, test_files, sanitised_dict):
 def test_collation_skips(tmp_path, test_files, uploads_dict):
     fileset = test_files["basic"]
     config = load_config(fileset["config_excel"])
-    matcher = Matcher(config)
+    matcher = Matcher(
+        config["paths"],
+        config["mappings"],
+        SPREADSHEET_FIELDS,
+        dicom_extractor,
+        XNATFileMatch,
+    )
     log = tmp_path / "log.xlsx"
     new_workbook(log)
     scan(matcher, Path(fileset["dir"]), log)
-    project_id = uploads_dict["project"]
     wb = load_workbook(log)
     ws = wb["Files"]
     header = True
@@ -110,7 +132,7 @@ def test_collation_skips(tmp_path, test_files, uploads_dict):
                     logger.warning(f"Skipping {matchfile.file}")
                     n += 1
             files.append(matchfile)
-    skip, uploads = collate_uploads(project_id, files)
+    skip, uploads = collate_uploads(files)
     assert len(skip) == uploads_dict["skipped"] + skip_n
     for session_scan, upload in uploads.items():
         uploads[session_scan] = [f.file for f in uploads[session_scan].files]
@@ -128,7 +150,13 @@ def test_collation_skips(tmp_path, test_files, uploads_dict):
 def test_secret_pdfs(tmp_path, test_files):
     fileset = test_files["secret_pdf"]
     config = load_config(fileset["config_excel"])
-    matcher = Matcher(config)
+    matcher = Matcher(
+        config["paths"],
+        config["mappings"],
+        SPREADSHEET_FIELDS,
+        dicom_extractor,
+        XNATFileMatch,
+    )
     scanned = tmp_path / "scanned.xlsx"
     new_workbook(scanned)
     scan(matcher, Path(fileset["dir"]), scanned, include_unmatched=True)